corrections in loginScreenContent, added AutoResizedText, extracted ThemeSelectorRow from ThemePreferences
