+++ app/src/main/java/com/savenko/track/data/other/constants/EntitiesDefaults.kt
@@ -11,7 +11,7 @@ const val NAME_DEFAULT: String = "User"
 const val LOGIN_COUNT_DEFAULT = 0
 const val BUDGET_DEFAULT: Float = 0.0f
 const val SHOW_PAGE_NAME_DEFAULT: Boolean = true
-const val USE_SYSTEM_THEME_DEFAULT: Boolean = true
+const val USE_SYSTEM_THEME_DEFAULT: Boolean = false
 const val NON_CATEGORY_FINANCIALS_DEFAULT = false
 const val GROUPING_CATEGORY_ID_DEFAULT = Int.MIN_VALUE // This value used to compare if user selected any category
 const val EXPENSE_CATEGORY_GROUPING_ID_DEFAULT = 13
