+++ app/src/main/java/com/savenko/track/presentation/screens/core/LoginScreen.kt
@@ -1,87 +1,35 @@
 package com.savenko.track.presentation.screens.core
 
 import android.app.Activity
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExposedDropdownMenuBox
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.paint
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.platform.LocalView
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.core.view.WindowCompat
 import androidx.navigation.NavController
 import com.savenko.track.R
-import com.savenko.track.data.other.constants.MAX_BUDGET_VALUE
-import com.savenko.track.data.other.constants.NAME_MAX_LENGTH
-import com.savenko.track.data.viewmodels.login.LoginViewModel
-import com.savenko.track.domain.models.currency.Currency
-import com.savenko.track.presentation.navigation.Screen
-import kotlinx.coroutines.launch
+import com.savenko.track.data.viewmodels.login.LoginScreenViewModel
+import com.savenko.track.presentation.screens.loginScreen.LoginHeader
+import com.savenko.track.presentation.screens.loginScreen.LoginScreenContent
 import org.koin.androidx.compose.koinViewModel
 
 
 @Composable
 fun LoginScreen(navController: NavController) {
     val localContext = LocalContext.current
-    val loginViewModel = koinViewModel<LoginViewModel>()
+    val loginScreenViewModel = koinViewModel<LoginScreenViewModel>()
     LaunchedEffect(key1 = Unit) {
-        loginViewModel.setFirstNameStateFlow(localContext.getString(R.string.user))
+        loginScreenViewModel.setFirstNameStateFlow(localContext.getString(R.string.user))
     }
     Surface(modifier = Modifier.fillMaxSize()) {
         val useDarkTheme = isSystemInDarkTheme()
@@ -99,257 +47,10 @@ fun LoginScreen(navController: NavController) {
             modifier = Modifier.fillMaxSize(),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            LoginHeader()
-            Spacer(modifier = Modifier.height(32.dp))
-            LoginContent(loginViewModel, navController)
+            LoginHeader()//LoginScreenHeader()
+            Spacer(modifier = Modifier.weight(1f))
+            LoginScreenContent(loginScreenViewModel, navController)
+            Spacer(modifier = Modifier.weight(1.2f))
         }
     }
 }
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun LoginContent(loginViewModel: LoginViewModel, navController: NavController) {
-    val focusManager = LocalFocusManager.current
-    val controller = LocalSoftwareKeyboardController.current
-    val nameFocusRequester = remember { FocusRequester() }
-    val budgetFocusRequester = remember { FocusRequester() }
-    val coroutineScope = rememberCoroutineScope()
-    val screenState = loginViewModel.loginScreenState.collectAsState()
-    Box(
-        modifier = Modifier
-            .heightIn(200.dp, Dp.Infinity)
-            .padding(horizontal = 8.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .wrapContentHeight()
-                .padding(horizontal = 8.dp, vertical = 8.dp)
-                .paint(
-                    painterResource(id = R.drawable.login_background),
-                    contentScale = ContentScale.FillBounds,
-                    sizeToIntrinsics = false,
-                    colorFilter = ColorFilter.tint(
-                        MaterialTheme.colorScheme.primaryContainer
-                    )
-                ),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.SpaceEvenly
-        ) {
-            FlowRow(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.Center,
-                verticalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = stringResource(R.string.name),
-                    style = MaterialTheme.typography.bodyMedium,
-                    textAlign = TextAlign.Center,
-                    color = MaterialTheme.colorScheme.onPrimaryContainer,
-                    modifier = Modifier.align(Alignment.CenterVertically)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                BasicTextField(
-                    modifier = Modifier
-                        .focusRequester(nameFocusRequester)
-                        .width(IntrinsicSize.Min)
-                        .padding(start = 8.dp)
-                        .align(Alignment.CenterVertically),
-                    textStyle = MaterialTheme.typography.titleLarge.copy(
-                        letterSpacing = 1.sp,
-                        color = MaterialTheme.colorScheme.onPrimaryContainer,
-                        fontWeight = FontWeight.W600
-                    ),
-                    value = screenState.value.name,
-                    onValueChange = { newText ->
-                        if (newText.length < NAME_MAX_LENGTH) {
-                            loginViewModel.setFirstNameStateFlow(newText)
-                        }
-                    },
-                    keyboardOptions = KeyboardOptions(
-                        keyboardType = KeyboardType.Text,
-                        imeAction = ImeAction.Done
-                    ),
-                    keyboardActions = KeyboardActions(
-                        onDone = {
-                            controller?.hide()
-                            focusManager.clearFocus()
-                        }
-                    )
-                )
-            }
-            FlowRow(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = stringResource(
-                        id = R.string.aprox_month_income_login_screen
-                    ),
-                    style = MaterialTheme.typography.bodyMedium, textAlign = TextAlign.Center,
-                    modifier = Modifier.align(Alignment.CenterVertically)
-                )
-                Row(modifier = Modifier.wrapContentWidth(), verticalAlignment = Alignment.CenterVertically) {
-                    BasicTextField(
-                        modifier = Modifier
-                            .focusRequester(budgetFocusRequester)
-                            .width(IntrinsicSize.Min)
-                            .padding(start = 12.dp),
-                        textStyle = MaterialTheme.typography.titleLarge.copy(
-                            color = MaterialTheme.colorScheme.onPrimaryContainer,
-                            fontWeight = FontWeight.W600,
-                            letterSpacing = (0.8).sp
-                        ),
-                        value = screenState.value.budget.toString(),
-                        onValueChange = { newText ->
-                            try {
-                                val value = newText.toFloat()
-                                if (value < MAX_BUDGET_VALUE) {
-                                    loginViewModel.setIncomeStateFlow(
-                                        value
-                                    )
-                                }
-                            } catch (e: NumberFormatException) {
-                                screenState.value.budget
-                            }
-                        },
-                        keyboardOptions = KeyboardOptions(
-                            keyboardType = KeyboardType.Number,
-                            imeAction = ImeAction.Done
-                        ),
-                        keyboardActions = KeyboardActions(
-                            onDone = {
-                                controller?.hide()
-                                focusManager.clearFocus()
-                            }
-                        )
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    LoginScreenCurrencyPicker(
-                        currencyList = loginViewModel.currencyList,
-                        selectedOption = screenState.value.currency,
-                        onSelect = {
-                            loginViewModel.setCurrencyStateFlow(it)
-                        }
-                    )
-                }
-
-            }
-            Button(
-                modifier = Modifier
-                    .wrapContentHeight()
-                    .wrapContentWidth(),
-                onClick = {
-                    coroutineScope.launch {
-                        loginViewModel.addToDataStore()
-                    }
-                    navController.navigate(Screen.MainScreen.route)
-                },
-                colors = ButtonDefaults.buttonColors(),
-                shape = RoundedCornerShape(16.dp)
-            ) {
-                Text(
-                    text = stringResource(R.string.lets_start),
-                    style = MaterialTheme.typography.bodyMedium.copy(fontWeight = FontWeight.Medium),
-                    modifier = Modifier.align(Alignment.CenterVertically)
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun LoginHeader() {
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        Image(
-            modifier = Modifier.size(64.dp),
-            painter = painterResource(id = R.drawable.track_new_icon),
-            contentDescription = stringResource(id = R.string.app_logo)
-        )
-        Spacer(modifier = Modifier.width(16.dp))
-        Column(
-            modifier = Modifier
-                .padding(vertical = 8.dp)
-                .wrapContentWidth()
-        ) {
-            Text(
-                text = stringResource(id = R.string.app_name),
-                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.W600),
-                color = MaterialTheme.colorScheme.onPrimary,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.align(Alignment.CenterHorizontally)
-            )
-            Text(
-                text = stringResource(id = R.string.logo_app_description),
-                style = MaterialTheme.typography.bodySmall.copy(fontWeight = FontWeight.W500),
-                textAlign = TextAlign.Center,
-                color = MaterialTheme.colorScheme.onPrimary
-            )
-        }
-    }
-
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-inline fun LoginScreenCurrencyPicker(
-    currencyList: List<Currency>,
-    selectedOption: Currency,
-    crossinline onSelect: (Currency) -> Unit
-) {
-    val uiColor = if (isSystemInDarkTheme()) Color.White else Color.Black
-    var isExpanded by remember { mutableStateOf(false) }
-    val focusRequester = remember { FocusRequester() }
-    val focusManager = LocalFocusManager.current
-
-    ExposedDropdownMenuBox(
-        expanded = isExpanded,
-        onExpandedChange = {
-            isExpanded = !isExpanded
-            if (isExpanded) {
-                focusRequester.requestFocus()
-            } else {
-                focusManager.clearFocus()
-            }
-        }, modifier = Modifier
-            .wrapContentWidth()
-            .focusRequester(focusRequester)
-    ) {
-        Card(
-            modifier = Modifier.menuAnchor(),
-            shape = RoundedCornerShape(8.dp)
-        ) {
-            BasicTextField(
-                value = selectedOption.ticker,
-                readOnly = true,
-                modifier = Modifier
-                    .width(IntrinsicSize.Min)
-                    .padding(8.dp),
-                onValueChange = {},
-                textStyle = MaterialTheme.typography.titleSmall.copy(color = MaterialTheme.colorScheme.primary)
-            )
-        }
-        ExposedDropdownMenu(
-            modifier = Modifier.width(IntrinsicSize.Min),
-            expanded = isExpanded,
-            onDismissRequest = {
-                isExpanded = false
-            }
-        ) {
-            currencyList.forEach { selectionOption ->
-                DropdownMenuItem(
-                    text = { Text(text = selectionOption.name, color = uiColor) },
-                    onClick = {
-                        onSelect(selectionOption)
-                        isExpanded = false
-                        focusManager.clearFocus()
-                    },
-                    trailingIcon = {
-                        Text(text = selectionOption.ticker, color = uiColor)
-                    }
-                )
-            }
-        }
-    }
-}
-
