+++ app/src/main/java/com/savenko/track/presentation/screens/loginScreen/LoginScreenHeader.kt
@@ -1,0 +1,121 @@
+package com.savenko.track.presentation.screens.loginScreen
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentWidth
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawWithContent
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.nativeCanvas
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.savenko.track.R
+
+@Composable
+fun LoginScreenHeader() {
+    Row(verticalAlignment = Alignment.CenterVertically) {
+        Image(
+            modifier = Modifier.size(64.dp),
+            painter = painterResource(id = R.drawable.track_new_icon),
+            contentDescription = stringResource(id = R.string.app_logo)
+        )
+        Spacer(modifier = Modifier.width(16.dp))
+        Column(
+            modifier = Modifier
+                .padding(vertical = 8.dp)
+                .wrapContentWidth()
+        ) {
+            Text(
+                text = stringResource(id = R.string.app_name),
+                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.W500),
+                textAlign = TextAlign.Center,
+                modifier = Modifier.align(Alignment.CenterHorizontally)
+            )
+            Text(
+                text = stringResource(id = R.string.logo_app_description),
+                style = MaterialTheme.typography.bodySmall.copy(fontWeight = FontWeight.W400),
+                textAlign = TextAlign.Center,
+            )
+        }
+    }
+}
+
+@Composable
+fun LoginHeader() {
+    val brush = Brush.linearGradient(
+        colors = listOf(
+            MaterialTheme.colorScheme.tertiary,
+            MaterialTheme.colorScheme.primary,
+        )
+    )
+    val vector = ImageVector.vectorResource(id = R.drawable.header)
+    Box() {
+        Icon(
+            imageVector = vector,
+            contentDescription = null,
+            modifier = Modifier
+                .align(Alignment.TopCenter)
+                .drawWithContent {
+                    with(drawContext.canvas.nativeCanvas) {
+                        val checkPoint = saveLayer(null, null)
+                        drawContent()
+                        drawRect(
+                            brush,
+                            topLeft = Offset(0f, 0f),
+                            size = Size(size.width, size.height),
+                            blendMode = BlendMode.SrcIn
+                        )
+                        restoreToCount(checkPoint)
+                    }
+                }
+        )
+        Box(modifier = Modifier.align(Alignment.TopCenter)) {
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                Image(
+                    modifier = Modifier.size(64.dp),
+                    painter = painterResource(id = R.drawable.track_new_icon),
+                    contentDescription = stringResource(id = R.string.app_logo)
+                )
+                Spacer(modifier = Modifier.width(8.dp))
+                Column(
+                    modifier = Modifier
+                        .padding(vertical = 8.dp)
+                        .wrapContentWidth()
+                ) {
+                    Text(
+                        text = stringResource(id = R.string.app_name),
+                        style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.W600),
+                        color = MaterialTheme.colorScheme.onPrimary,
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier.align(Alignment.CenterHorizontally)
+                    )
+                    Text(
+                        text = stringResource(id = R.string.logo_app_description),
+                        style = MaterialTheme.typography.bodySmall.copy(fontWeight = FontWeight.W500),
+                        textAlign = TextAlign.Center,
+                        color = MaterialTheme.colorScheme.onPrimary
+                    )
+                }
+            }
+        }
+    }
+}
